style: 파일명과 컴포넌트명 변경
